Fix user preference code error
Fix invalid UserPreference creation - was using incorrect field.

Backfill user preference records instead of doing on first user-fetch

Upgrade prisma

remove IconLazyWrapper to avoid layout shift
