+++ libs/ui/src/lib/widgets/Icon.tsx
@@ -1,29 +1,33 @@
-// https://www.lightningdesignsystem.com/components/icons/
-
-import { IconName, IconType } from '@jetstream/icon-factory';
-import React, { FunctionComponent, Suspense } from 'react';
-
-const IconLazyWrapper = React.lazy(() => import('./IconLazyWrapper'));
+import { IconName, IconType, getIcon } from '@jetstream/icon-factory';
+import classNames from 'classnames';
 
 export interface IconProps {
-  containerClassname?: string; // container classname, only used if not omitted
-  className?: string; // SVG element classname
+  containerClassname?: string;
+  className?: string;
   omitContainer?: boolean;
   title?: string;
   type: IconType;
   icon: IconName;
   description?: string;
 }
+
 /**
- * This component lazy loads the actual icon class to allow code splitting
- * to reduce the size of the main bundle
+ * https://www.lightningdesignsystem.com/components/icons/
  */
-export const Icon: FunctionComponent<IconProps> = (props) => {
-  return (
-    <Suspense fallback={<svg />}>
-      <IconLazyWrapper {...props} />
-    </Suspense>
-  );
+export const Icon = ({ containerClassname, className, title, omitContainer = false, type, icon, description }: IconProps) => {
+  containerClassname = containerClassname || '';
+  className = className || '';
+  const svg = getIcon(type, icon, className);
+  if (omitContainer) {
+    return svg;
+  } else {
+    return (
+      <span className={classNames('slds-icon_container', containerClassname)} title={title || description}>
+        {svg}
+        {(description || title) && <span className="slds-assistive-text">{description || title}</span>}
+      </span>
+    );
+  }
 };
 
 export default Icon;
