+++ package.json
@@ -215,7 +215,7 @@
     "postcss": "8.4.19",
     "postcss-preset-env": "7",
     "prettier": "2.7.0",
-    "prisma": "^5.6.0",
+    "prisma": "^5.9.1",
     "react-refresh": "^0.10.0",
     "release-it": "^15.6.0",
     "sass": "1.55.0",
@@ -253,7 +253,7 @@
     "@opentelemetry/exporter-trace-otlp-grpc": "^0.27.0",
     "@opentelemetry/sdk-node": "^0.27.0",
     "@popperjs/core": "^2.4.0",
-    "@prisma/client": "^5.6.0",
+    "@prisma/client": "^5.9.1",
     "@react-aria/dialog": "^3.1.4",
     "@react-aria/focus": "^3.5.0",
     "@react-aria/overlays": "^3.7.2",
